컨포넌트 분리 및 코드 리펙토링
