+++ package.json
@@ -10,9 +10,12 @@
     "@types/node": "^16.11.21",
     "@types/react": "^17.0.38",
     "@types/react-dom": "^17.0.11",
+    "@types/styled-components": "^5.1.22",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
     "react-scripts": "5.0.0",
+    "react-spinners": "^0.11.0",
+    "styled-components": "^5.3.3",
     "typescript": "^4.5.5"
   },
   "scripts": {
