+++ src/index.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import ReactDOM from "react-dom";
-import App from "./App";
+import App from "./pages/App";
 
 ReactDOM.render(
   <React.StrictMode>
