+++ src/types/interface.ts
@@ -1,3 +1,13 @@
+export const initialState = {
+  basePrice: 0,
+  openingPrice: 0,
+  changePrice: 0,
+  cashBuyingPrice: 0,
+  cashSellingPrice: 0,
+  ttSellingPrice: 0,
+  ttBuyingPrice: 0,
+};
+
 export interface data {
   basePrice: number;
   openingPrice: number;
